feat :connect to github

+++ app/env/env.sh
@@ -1,16 +1,42 @@
 #!/bin/bash
 
-function hello_world(name) {
-  echo "Hello, $name!"
+hello_world() {
+    read -p "Enter your name : " name
+    echo "Hello, $name!"
 }
 
-function run(){
+run(){
     echo "Welcome in BackupLik"
 }
 
-function installGit(){
-    sudo apt update
-    sudo apt install git
+installGit(){
+    if [ -x "$(command -v git)" ]; then
+        echo "Git is already installed"
+    else
+        echo "Git is not installed"
+        echo "Do you want to install git ? (y/n)"
+        read response
+        if [ $response == "y" ]; then
+            sudo apt update
+            sudo apt upgrade
+            sudo apt install git
+        else
+            echo "Git is required to run this script"
+            exit 1
+        fi
+    fi
+
+    if [ ! -x "$(command -v xdg-open)" ]; then
+        echo "xdg-open is not installed"
+        echo "Do you want to install xdg-open ? (y/n)"
+        read response
+        if [ $response == "y" ]; then
+            sudo apt install xdg-utils
+        else
+            echo "xdg-open is required to run this script"
+            exit 1
+        fi
+    fi
 
     echo "Entre your github name : "
     read name
@@ -18,16 +44,37 @@ function installGit(){
     echo "Entre your github mail : "
     read mail
 
-    ssh-keygen -t rsa -b 4096 -C email
-    cd /home/$USER/.ssh 
+    git config --global user.name $name
+    git config --global user.email $mail
+
+    echo "Git is installed"
+    echo `git config --global --list`
 }
 
-function fileBackup(){
+sshConnect(){
+    echo "Generate ssh key"
+    echo "Entre your github mail : "
+    read email
+    ssh-keygen -t rsa -b 4096 -C $email
+    cat /$USER/.ssh/id_rsa.pub
+    echo "Add this key to your github account"
+    xdg-open https://github.com/settings/keys
+
+    echo "Entre your github ssh key: "      
+    xdg-open https://github.com/settings/tokens
+    read key
+
+    git init
+    git remote -v
+
+}
+
+fileBackup(){
     echo "Entre Path of file : " #path relatif / absolute
     read path
 }
 
-function connectToGithub(name,mail){
+connectToGithub(){
     git config --global user.name name
     git config --global user.email email
 
@@ -38,3 +85,7 @@ function connectToGithub(name,mail){
     git remote -v
 
 }
+
+# hello_world
+# installGit
+sshConnect
